client, store: moved children responsibility
Before, the put handler in the client waited for the children to be available,
preventing the handler from finishing immediately. Now this is done in the store,
i.e. in the collection task of the client. With this, there is no need for each
handler to be its own task anymore.

This is a preparation for #36.

Also, we do not handle yet fetching multipart resources over the network in
workers, where a similar logic would be necessary, but that's also not a use
case we ever encountered.