tests: adjusted timeouts
Now that the client has a tunable retry rate, some tests can take a bit more
time, so made the timeout more permissive.

The current timeouts are enough currently with a retry delay of 0.1s to 1s, on
my machine, and take respectively 30s and 1m to go through the test suite.