refactor commands: idempotence, replies
 * Made the creation of both get and rget idempotent
 * This in turn enables a much lighter style when defining rget and collect
 * Move the out argument to a local deque

The local deque has a little side-effect programming style, especially as it's
easy to forget to flush it. But it's not worse by itself than returning replies
and ignoring them, which could drop messages. In the future we'll use commands
everywhere, so the loop would have a check for it. At this point risk of not
flushing the command queue will be void, and risks of functions silently
discarding new commands will also be avoided.

But for now remember to check or flush the new_command queue if you set command
flags.

Riders:
 * the debug config option was turned into a more general log_level