injectable path maker prototype
Steps can request a `_new_path` input, which will be injected on worker-side and
generates unique paths for each (task, call sequence number) combination.

This solves the naming problem when a step wants to transform a file into an
other file and has to generate a reproducible, unique name to store the result.