feat(client, worker): recursive tasks prototype
Draft implementation of the recursive task idea:
 * the serializer returns the list of Tasks encountered when processing a task's
   native result,
 * the worker serializes this list (and the whole sub-graph !) and appends it to
   the DONE message,
 * the client deserializes the new graph and adds it to its own
 * the client injects it as result of the "submit" command
 * the rsubmit starts the recursive resubmits based on it

For this proof of concept we serialize brutally with dill, but we'll switch to
something more python-agnostic at some point.

Almost as much log reformatting than code to implement the future to make the
info log more useful when developping this kind of thing.