feat(galp): meta tasks, with task persistency
Implemented the STAT/FOUND message pair to allow a client to fetch task
definitions from the worker's storage. Some code sections changed to accept just
name or task dictionaries instead of task object.

Use of task objects should be phased out of most of the code, for now some code
sections need to work with either one or the other.