feat(graph): true injection system
Allows arguments to be used to provide injected args by delaying task creation.

There are still certainly combinations of features that do not work, consider
this a proof of concept (especially, literals inside the graph, worker
injects deeper inside the graph are not yet handled correctly, and behaviour
when used with nested tasks is still to be even defined).