refactor(graph): start epuring data model
We started with objects that had only few fields exposed over the network. As
the use cases became more complicated, we started to generate dictionaries of
basic types serializable through msgpack or json for ipc. This created a
redundancy between the original objects, and a loose collection of dictionaries
and helper functions that duplicated some of the model.

With this patch series, we introduce actual object models with automatic dumping
and validation, simplifying and defining more strictly our model.