refactor(graph): Major data model refactor, see below
We used to have some python objects representing a graph of computations, with
some select attributes exported over the network as bytes objects.

Over time, we started to send larger and larger part of this graph over the
wire. With meta-steps, especially, graph components can be completely remotely
created and discovered over the network by the client.

To allow this, we had to emit whole dictionaries of data, and introduced msgpack
as a serialization method. Many parts of the code had to work with such
dictionaries received over the network instead of the original objects.

At this point, we had an essentially duplicated object model, with the original
design and a flurry of badly defined dictionaries duplicating the information.

This patch series cleans up this state by rewriting the object model around
Pydantic models. These have an automatic, well-defined way to be converted to
dictionaries and from there msgpack buffers, and most importantly provides a way
to validate such data and reconstruct native objects.

Therefore, large parts of the code can now work again on validated, strongly
typed, native objects again. The duplication of the object model is eliminated,
and a lot of the code is now correctly typed.