fix(task_type): distinguish NamedCoreTaskDef
Some code paths, especially everythin related to task submission and execution,
work with named defs but expect them to be core named defs. Without a separate
type, there was no way to type this, and we had to add spurious asserts to type
check.

This makes named defs a generic, and adds aliases, type guards and validators
for both broad and core named defs.