refactor(protocol): simplify submit message format
Now only a serialized named task is sent over the network, without the redundant
name. More validation added to this and to TaskName.

The rationale is that we want to eventually get rid of all the custom message
building and validation, and instead levae that completely to pydantic/msgpack.
This is an experiment in this direction.

Next steps would be to do the same for all messages  that have several data
frames, and finally to only send a "Galp message" object without any
serialization/validation ourselves at all.