feat(broker, pool): on-demand worker creation
Prepare for actual resource allocation by having the broker explicitly track
resources, and the pool spin up workers as requested by the pool.

This entailed several refactors. Broker was considerably simplified by the
observation that there was in fact no need for separate endpoints for workers
and clients ; they are now fused.

Furthermore, we started to write a model of galp messages, starting humbly with
Ready.