fix(broker): avoid multiple allocation of same tasks
This was normally done by sending 'DOING' messages to tell clients to stop
retrying. Now that workers are spawned on demand, a request may be re-sent many
times before the first DOING gets through, so this does not work anymore.

Instead, two solutions should be implemented:
 * This commit: the broker should check incoming tasks against the allocations
   that it already tracks and optimize out double submissions
 * Todo: the broker should already send a DOING message it self to stop the
   client from retrying during the worker start phase

Also the structure of the parallel test was modified to be less flacky under the
new startup system.