refactor(protocol): split message creation and routing
Now messages can be created without a Route, and a `route_message` callback can
be used to fill in the route afterwars. For clients and workers the routing
logic is quite straightforward so that simplifies the code, for brokers the new
intermediate-level interface is also much more practical.

On the other hand, places where a message needs to be manually sent get slighlty more
verbose.