refactor(commands): functional primitive collection
The existing code would store new primitives in a script-global list at
creation, which had to be checked in a separate step. This in turns means that
functional command creation is impossible has we require a reference to that
list at all time.

Instead, make sure new primitives are functionally passed down.