refactor(commands): avoid recursing on non-pending commands
With class-based commands, it was necessary to initialize each command to obtain
the list of events to proapagate. Now that we got rid of these, it is only
necessary to inspect the terminal commands (primitives and empty gathers).
Every change to the command graph can then be deduced by propagating any change
of state upstream.

Collecting leaves should be done on any new commands, but not on the inputs of
advance_all which should be pending but maybe ready commands, this was a badly
located fix.