refactor(commands, worker): simplify submission workflow
Now that we propagate task definitions through the command system, we can
directly use them to generate submit messages without relying on the client
having a comprehensive repository of task defs.

However, we currently still need a repository of _local_ tasks.
 * We rely on it for double add guard, though this is an expendable feature.
 * More critically, we rely on it for cases like 20/structured_inputs, probably
   for things like tasks nested inside literals or something of that sort.

That being said, the tracking of remote definitions, on the other hand, is now
unneeded and could be safely scrapped.