refactor: move trivial uses of route_message
The Protocol, which is a handler case, wrongly exposes some sending routines.
Many uses of such routines is to send trivially addressed messages that require
no information from the protocol object, but are dependent on the stack. Move
that trivial functionality to the stack, which does have a sending purpose too,
and remove many accesses to it through the protocol object.