refactor: introduce UpperForwardingSession
Recent refactor attempts uncovered that the original session interface design
was not enough for case where the decision to forward a message to a peer would
be done based on app logic. This commit introduces a richer session interface
that should cover such cases. A first use of this was done in the broker as a
proof of concept that this can replace the RoutedMessage interface.