refactor: switch from routes to sessions in broker
This eliminates the last broker location where a RoutedMessage was created de
novo. Since the broker tracks per-worker information, this required still
exposing the route through an uid interface. However, the uid is opaque, and
this allows us, for, instance, to correctly encapsulate the fact that the route
need to be converted into a tuple to be hashable and used as a key.