refactor(client): simplify task status to a boolean
The role of marking which commands are done had been moved to the `is_pending`
flagged of the promises, and is what is actually checked to decide to requeue or
drop a command.

Therfore, the only place still using task statuses is to force early dropping
when we receive a DOING. This commit makes that obvious by replacing it by a
boolean that is only set in the doing handler.