refactor: do not assume return type of handlers
The type dispatcher that we wrote was only compatible with "immediate" handlers
that return a galp message to be passed to upper session. But that may not be
true of all handlers, and should be an implementation-wise choice.

Instead, assume handlers have a generic return type: give them a convenient
message writer, but let them write messages with anything they want ; and
provide utils to bridge the gap with simpler handlers.