refactor: improve types and names for sessions and handlers.
Now that we inject the session creation in the lower layer, session types, and
the types of the handler accepting these sessions, are generic in the routing
layer. On the other hand, the upper layer should only ever care about the
parametrized generic version of these types. Therefore, name and type alias
things more clearly.