refactor(*): Introduce Upload request
Major protocol change. Now, when the client encounters a Literal in the graph,
it sends an Upload Request to put them in the store. Then, the next task
submission that depends on this literal will directly find it in store instead
of asking the client.

The old functionality is still there, further testing needed before removing it.