refactor: unify Put/Serialized types
With this, the type returned/accepted by the store, and send/received in
Get/Upload exchanges, retuned by serializer and stored in Literal TaskNodes is
now the same, the Serialized type in task_types. This remove a lot of
conversions between store types and Put ; and avoid extra serialization
occuring.

The hierarchy of the Serializer class is still too complicated, but that's for
later.

Immediate benefit is simplification of e.g. Upload code. Objective is too allow
also shortcut get/rget in command with literal tasks, for now this still relies
on the client store.