Simplify primitive tracking
We actually do not care about caching settled primitive values, we only want to
avoid sending a request if an identical one is in flight. Therefore, we don't
need to track primitive values at all. Instead we can check the list of
instances directly to know if some are in flight. This requires determinisic
handling of instances, though, so remove the WeakSet mechanism.

This means we have strong references to all commands for which we still expect a
reply, which could theoretically leak memory until we implement somethings like
timeouts that clear pending commands.