Expose richer asyn interface for error handling, close #83
The new interface is still somewhat bad, because we use a boolean flag to
control whether errors should cause a whole collection to fail, or be returned
inside the list of results, therefore we have a type that change depending on
the flag value and need an overcomplicated return type. In time we should also
drop that flag and just have separate gather primitives with more accurate
signatures.