Move tests out of package tree
The resulting layout is still not optimal, because to get imports to work during
testing, we need to rely on pytest including the base dir in the sys.path when
tests/ contains an __init__.py and that makes the local copy of galp importable.

But our test setup is somewhat complex (we have config files and python modules
that need to be included in both the tests modules and the workers, as plug-ins)
so getting to work at all is already good for now.