Drop ErrT type parameter from Result
If you want to specify an error type, it's clearer and shorter to just write
the union explicitely. If you don't, it's much much better to have Result
hardcode Error, because it's shorter and crucially do not default to Any if you
omit the error type. Having Result = Any when you forget the parameters
effectively nullified the checking of result, and was why the inconsistency in
filter_commands was not flagged by the type checker.