Drop ErrT error type parameter from Command
This is similar to the last patch for Result, except in that case we don't have
a replacement. From now on, Command cannot be typed to specify that they can
only return specific errors, as this is a feature that we never used but forced us
to add a lot of extra Error or implicit Any in all the other cases.