feat(asyn): implement lifelong slot semantics
This means that a command only ever get one Slot assigned to it, and re-using a
command guarantees the same Slot will be used, meaning any progress done
previously on that command is preserved.

This complicates some logic slightly as we need to watch out for new command
whose inputs are already available.