refactor(galp): remove profiling and config file
Profiling tests were broken in recent commit, and was already flagged for
removal. Now that we have local reruns that allow users to fulfill their
profiling needs better that the built-in profiling, it's a good time to remove
it.

Since this was the last use case for the config file, also remove the latter.