refactor(pool): switch from zmq to unix socket
This because for communication with a forkserver, you clearly need a transport
that is safe to fork; however ZMQ is not; you're only supposed to create the zmq
context when you're sure you won't fork. So to ask a forkserver to fork a pipe
or unix socket is safer. With a datagram socket we don't need to worry about
message framing which helps a lot; we only lose the ability to send what zmq
calls multipart messages but that's easy enough to reimplement later if needed.