feature(galp): std stream forwarding to client
This is still mostly a demonstration. If verbose is off (default), this does
"pass-through" logging, just printing out as received. This emulates local
execution, which is redundant on a local machine as streams are shared, but
should become nice to have once we move to remote machines. If verbose is true,
this has the added value of tagging each chunk of input, which may turn out to
be a nice feature.

This commit mostly implement the basic infrastructure; the exact interface,
behavior and implementation details are left for later.