refactor(galp): split and remerge task types
Basic task types are moved to task_defs to shorten task_types, making room to
merge graph into task_types as Tasks and Steps really need to be in one module
because of the natural cyclic references.

This solves a longstanding cyclic import that worked with some hacks until know
and made linters complain.