fix(logserver): fix two fd leaks
The worker was not closing the pipes, as a result the forkserver also kept them
open. The fork handler to close pipes was not correct in case of already closed
pipes, but this was masked by the pipe never being closed. The worker also
wasn't closing the saved fds correctly.

While no test is added to the test suite, this is actually testable with a bit
of patience as leaking two fds per task typically results in crashes after 512
tasks have been run for systems where the limit is 1024.