refactor(broker): high-level forwarding
Until now, we were using the first frames in the messages to store origin
information for requests, which we then used to route back the responses
correctly. Starting with the logserver introduction, we started to use the
request ids and the broker state to dynamically deduce where to forward. Switch
to the latter solution for all replies, and stop using the low-level forward
solution.